requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
pydantic==2.7.4
jinja2==3.1.4

